Add as.list() method for correlation objects
There was a request on twitter for a list like output of c, N, p matrices. This produces such a list structure, with flexibility for expansion or customization by users.

I haven't done all of the md/html/etc. print methods. I'm also not sure how to do the p value formatting for the p matrix.